requirements.txt
python-telegram-bot>=20.0
mysql-connector-python>=8.0
python-dotenv>=1.0
pytest>=7.0
pytest-asyncio>=0.21
